requirements.txt
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==38.0.4
Flask==2.3.3
Flask-JWT-Extended==4.6.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==0.21.1
setuptools==68.2.2
Werkzeug==3.0.2
wheel==0.41.2
